requirements.txt
streamlit
gradio
pandas
google-generativeai
PyPDF2